backend/app/services/model_service.py
# File contains logic to load the machine learning models and make predictions
import joblib

model = joblib.load('path/to/your/model.joblib')

def predict_model(input_data):
    # Assuming the input data is already validates and preprocessed
    features = [[input_data.feature1, input_data.feature2]]
    prediction = model.predict(features)
    return prediction[0]