backend/app/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==2.0.0
sqlalchemy==2.0.17
asyncpg==0.27.0
databases==0.7.2
psycopg2-binary==2.9.6
alembic==1.12.2
httpx==0.24.1
python-dotenv==1.0.0
